// src/app/types.ts
type Trecord = {
    ID: number;
    Nome_Gioco: string;
    Anno: number;
    Piattaforma: string;
    Genere: string;
    Meccanica: string;
    Dinamica: string;
    Estetica: string;
    Sviluppo: string;
    Pubblicazione: string;
    citato_in: string;
    Stagione: number;
    Episodio: number;
    Note: string;
    Link: string;

};

export type { Trecord }